SP2-90 Refactor CsvHandle into Heatington.Utility namespace
The CsvHandle namespace and its associated classes have been refactored into the Heatington.Utility namespace. The CsvHandle.CsvConstructorAttribute has been renamed to CsvConstructor. Dependents have been adjusted accordingly for this change.
